Added horrific rule classification system.
Apologies in advance to those who read RuleStringRepository.java.

The current game rules are now loaded in such a way that agents can reason about them based on preset classification values.

Manipulated time stamping of actions to line up properly with the logical steps in the simulation's game of Nomic.