Altered NomicAgent to give structure to proposals.
NomicAgent now contains an overloadable function, chooseProposal(), which is intended as an entry point for inherited classes, that have specified AI logic for making decisions. By default it returns null, indicating that the agent has no rule change proposal to make at the given time.

Any proposal returned by chooseProposal() (in some child class) will be properly registered with the environment, Drools, the NomicService, and the game of Nomic itself within the simulation.